COMP: Update Linux OpenCL integration retrieving clGetGLContextInfoKHR address
This commit is done anticipating the transition to support recent
implementation of the OpenCL library requiring to get the handle
of `clGetGLContextInfoKHR` function by address.

Since the function `clGetGLContextInfoKHR` is not listed in the list
of exported symbol associated with the OpenCL library built as part of
the `OpenCL-ICD-Loader` project[^1], this commit switches to using the
function `clGetExtensionFunctionAddressForPlatform`.

This consolidates the approach to be similar with what is done
for the Windows platform and is consistent with the description
of the standard [^2][^3]:

> [...] implementations may choose to also export those functions statically
> from the object libraries implementing those functions. However, portable
> applications cannot rely on this behavior.

[^1]: https://github.com/KhronosGroup/OpenCL-ICD-Loader/blob/v2022.09.30/loader/linux/icd_exports.map
[^2]: https://registry.khronos.org/OpenCL/sdk/1.0/docs/man/xhtml/clGetExtensionFunctionAddress.html
[^3]: https://registry.khronos.org/OpenCL/sdk/3.0/docs/man/html/clGetExtensionFunctionAddressForPlatform.html