COMP: Add FindFilesystem CMake module to detect required filesystem libraries
This commit introduces the FindFilesystem CMake module (copied from [1]), which
can detect if an additional library is required for proper filesystem functionality
(such as stdc++fs or c++fs).

Additionally, the commit updates the filesystem_compat module to take advantage
of the detected headers, simplifying the filesystem_compat header and ensuring
that the correct libraries are linked against.

[1] https://github.com/vector-of-bool/CMakeCM/blob/master/modules/FindFilesystem.cmake