ENH: Update protocol to accept heuristic method for optimization initialization
This commit introduces a backward incompatible changes requiring
an updated client.

The "optimizeFrame" command is updated to require an additional
field called "opt_init_heuristic" allowing to control the optimization
process by setting which heuristic should be used for initialization.

Accepted values are the following:
- CURRENT_FRAME
- PREVIOUS_FRAME
- LINEAR_EXTRAPOLATION
- SPLINE_INTERPOLATION

Note that the hard-coded behavior in trackDialog corresponding to
"previous frame" is removed from both the python and Matlab client.

To support the transition of existing scripts using the python or Matlab
client, the default value for "opt_init_heuristic" value is still set to
"previous frame".

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>