STYLE: Reduce length of Python docstrings and comments
This changes is done anticipating the integration of linting using ruff.

In connect.py, length is reduced to be < 120 (as specified in pyproject.toml)

In conf.py, length is reduced to be < 88 because length settings in
pyproject.html is not considered for files that are not in the same folder
and in a descendant folder.