ENH: Generalize background image calculations to support 16/32-bit radiographs (#245)
Refactor the working 8-bit background image calculation into a generic function
to work on 8, 16,  and 32-bit radiograph images. 
Previously, any radiograph that was not 8-bit would use the same broken 16-bit
method.  This 16-bit method failed to correctly loop over the image because the
loop failed to take the size  of the datatype into account. This bug was originally
introduced in commit 9bbce52 ("background mask 1", 2018-04-03).

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>