BUG: Fix out of bounds NCC value in viewport clipping
Regression introduced in 4704f9b ("BUG: Clip bounds of DRR projection to the
bounds of the radiograph", 2023-11-01).

Previously whenever a volume was outside the bounds of a radiograph image the
NCC value would be appended as 0. In this case it is possible that the PSO could
try to move the volume to be outside the bounds of all the radiographs.