ENH: Add mesh file field in new trial UI
Since the mesh file associated with each of the volumes in a trial are
required only for collision detection, the addition of the UI field is
done dynamically if compiled with Autoscoper_COLLISION_DETECTION. If in
the future the path to the mesh file is needed regardless, it can simply
be added statically to the `VolumeBox.ui` file and the added class
members from this commit can be removed.
This commit also fixes a small bug with relative vs. absolute paths when
parsing mesh file paths in a trial config. Future commits will establish
a new config version to specify these paths and will not compile this
logic when Autoscoper_COLLISION_DETECTION is off.