ENH: Add mesh file field in new trial UI
Since the mesh file associated with each of the volumes in a trial are
required only for collision detection, the UI field is hidden if
compiled without Autoscoper_COLLISION_DETECTION.
This commit also fixes a small bug with relative vs. absolute paths when
parsing mesh file paths in a trial config. Future commits will establish
a new config version to specify these paths and will not compile this
logic when Autoscoper_COLLISION_DETECTION is off.